Remove unused unimodules dependency
Unimodules are referenced here and there in the build process when setup. Since it's not the case here, I assume this module is not used (also it has a weird version, if a mistake or breaking change gets pushed to master on the github repo of unimodules it would instantly break our builds).